requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pillow==10.4.0
streamlit==1.37.1
requests==2.32.3
